Split up usage_plotter.py into separate modules - Update VSCode Jupyter working dir
