+++ .env.template
@@ -1,2 +1,2 @@
-#  Path to Apache access logs
-LOGS_PATH=../access_logs
+#  Path to Apache access logs (relative to root)
+LOGS_PATH=
