+++ .vscode/settings.json
@@ -15,5 +15,6 @@
   "python.testing.pytestArgs": [],
   "python.testing.unittestEnabled": false,
   "python.testing.nosetestsEnabled": false,
-  "python.testing.pytestEnabled": true
+  "python.testing.pytestEnabled": true,
+  "jupyter.notebookFileRoot": "${workspaceFolder}"
 }
